requirements.txt
pyaudio
speechrecognition
PyYAML
nltk
azure-cognitiveservices-speech